add CRUD for categories and user

+++ resources/views/dashboard/categories/list.blade.php
@@ -16,9 +16,9 @@
         <tr>
             <td>{{++$key}}</td>
             <td>{{$category->name}}</td>
-            <td><a type="button" class="btn btn-primary" href="/admin/categories/details/{{$category->id}}">Chi tiết</a>
+            <td>
                 <a type="button" class="btn btn-outline-warning" href="/admin/categories/edit/{{$category->id}}">Sửa</a>
-                <a type="button" class="btn btn-outline-danger" href="/admin/categories/delete/{{$category->id}}">Xóa</a>
+                <a type="button" class="btn btn-outline-danger" onClick="return confirm('Xoá thể loại {{$category->name}}?')" href="/admin/categories/delete/{{$category->id}}">Xóa</a>
             </td>
         </tr>
         @endforeach
